refactor: extract shared test helpers and fix userId assertion
Moved registerAndLogin and TestError into TestHelpers.swift, removing
three identical private copies across the category test suites.
Fixed the login success test to assert response.userId equals the known
user's id from the DB rather than comparing against a random UUID().
