+++ Tests/GrowBitAppServerTests/GrowBitAppServerSavingCategoryTests.swift
@@ -14,26 +14,6 @@ import Fluent
 @Suite("Category Creation Tests")
 struct GrowBitAppServerSavingCategoryTests {
 
-    private func registerAndLogin(in app: Application, username: String) async throws -> (token: String, userId: UUID) {
-        let user = User(username: username, password: "password")
-        try await app.testing().test(.POST, "/api/register") { req in
-            try req.content.encode(user)
-        } afterResponse: { res in
-            #expect(res.status == .ok)
-        }
-
-        var token = ""
-        var userId = UUID()
-        try await app.testing().test(.POST, "/api/login") { req in
-            try req.content.encode(user)
-        } afterResponse: { res in
-            let response = try res.content.decode(AuthResponseDTO.self)
-            token = response.token
-            userId = response.userId
-        }
-        return (token, userId)
-    }
-
     @Test("Category creation - Success")
     func categoryCreationSuccess() async throws {
         try await withApp(configure: configure) { app in
@@ -190,6 +170,3 @@ struct GrowBitAppServerSavingCategoryTests {
     }
 }
 
-enum TestError: Error {
-    case userCreationFailed
-}
